refactor: CardInfoHandler 코드 분리
