+++ src/app/context/cardInfo/index.ts
@@ -1,0 +1,1 @@
+export { CardInfoProvider } from './CardInfoProvider';
